+++ tests/Runtime/LambdaRuntimeTest.php
@@ -15,11 +15,10 @@ use Bref\Event\Sqs\SqsEvent;
 use Bref\Event\Sqs\SqsHandler;
 use Bref\Runtime\LambdaRuntime;
 use Bref\Runtime\ResponseTooBig;
+use Bref\Test\RuntimeTestCase;
 use Bref\Test\Server;
 use Exception;
 use GuzzleHttp\Psr7\Response;
-use JsonException;
-use PHPUnit\Framework\TestCase;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
@@ -30,21 +29,15 @@ use RuntimeException;
  *
  * The API is mocked using a fake HTTP server.
  */
-class LambdaRuntimeTest extends TestCase
+class LambdaRuntimeTest extends RuntimeTestCase
 {
     private LambdaRuntime $runtime;
 
     protected function setUp(): void
     {
-        ob_start();
-        Server::start();
-        $this->runtime = new LambdaRuntime('localhost:8126', 'phpunit');
-    }
+        parent::setUp();
 
-    protected function tearDown(): void
-    {
-        Server::stop();
-        ob_end_clean();
+        $this->runtime = new LambdaRuntime('localhost:8126', 'phpunit');
     }
 
     public function test basic behavior()
@@ -367,105 +360,5 @@ ERROR;
         $this->runtime->processNextEvent($handler);
 
         $this->assertEquals(new EventBridgeEvent($eventData), $handler->event);
-    }
-
-    private function givenAnEvent(mixed $event): void
-    {
-        Server::enqueue([
-            new Response( // lambda event
-                200,
-                [
-                    'lambda-runtime-aws-request-id' => '1',
-                    'lambda-runtime-invoked-function-arn' => 'test-function-name',
-                ],
-                json_encode($event, JSON_THROW_ON_ERROR)
-            ),
-            new Response(200), // lambda response accepted
-        ]);
-    }
-
-    private function assertInvocationResult(mixed $result)
-    {
-        $requests = Server::received();
-        $this->assertCount(2, $requests);
-
-        [$eventRequest, $eventResponse] = $requests;
-        $this->assertSame('GET', $eventRequest->getMethod());
-        $this->assertSame('http://localhost:8126/2018-06-01/runtime/invocation/next', $eventRequest->getUri()->__toString());
-        $this->assertSame('POST', $eventResponse->getMethod());
-        $this->assertSame('http://localhost:8126/2018-06-01/runtime/invocation/1/response', $eventResponse->getUri()->__toString());
-        $this->assertEquals($result, json_decode($eventResponse->getBody()->__toString(), true, 512, JSON_THROW_ON_ERROR));
-    }
-
-    private function assertInvocationErrorResult(string $errorClass, string $errorMessage)
-    {
-        $requests = Server::received();
-        $this->assertCount(2, $requests);
-
-        [$eventRequest, $eventResponse] = $requests;
-        $this->assertSame('GET', $eventRequest->getMethod());
-        $this->assertSame('http://localhost:8126/2018-06-01/runtime/invocation/next', $eventRequest->getUri()->__toString());
-        $this->assertSame('POST', $eventResponse->getMethod());
-        $this->assertSame('http://localhost:8126/2018-06-01/runtime/invocation/1/error', $eventResponse->getUri()->__toString());
-
-        // Check the content of the result of the lambda
-        $invocationResult = json_decode($eventResponse->getBody()->__toString(), true, 512, JSON_THROW_ON_ERROR);
-        $this->assertSame([
-            'errorType',
-            'errorMessage',
-            'stackTrace',
-        ], array_keys($invocationResult));
-        $this->assertEquals($errorClass, $invocationResult['errorType']);
-        $this->assertEquals($errorMessage, $invocationResult['errorMessage']);
-        $this->assertIsArray($invocationResult['stackTrace']);
-    }
-
-    private function assertErrorInLogs(string $errorClass, string $errorMessage): void
-    {
-        // Decode the logs from stdout
-        $stdout = $this->getActualOutput();
-
-        [$requestId, $message, $json] = explode("\t", $stdout);
-
-        $this->assertSame('Invoke Error', $message);
-
-        // Check the request ID matches a UUID
-        $this->assertNotEmpty($requestId);
-
-        try {
-            $invocationResult = json_decode($json, true, 512, JSON_THROW_ON_ERROR);
-        } catch (JsonException $e) {
-            $this->fail("Could not decode JSON from logs ({$e->getMessage()}): $json");
-        }
-        unset($invocationResult['previous']);
-        $this->assertSame([
-            'errorType',
-            'errorMessage',
-            'stack',
-        ], array_keys($invocationResult));
-        $this->assertEquals($errorClass, $invocationResult['errorType']);
-        $this->assertStringContainsString($errorMessage, $invocationResult['errorMessage']);
-        $this->assertIsArray($invocationResult['stack']);
-    }
-
-    private function assertPreviousErrorsInLogs(array $previousErrors)
-    {
-        // Decode the logs from stdout
-        $stdout = $this->getActualOutput();
-
-        [, , $json] = explode("\t", $stdout);
-
-        ['previous' => $previous] = json_decode($json, true, 512, JSON_THROW_ON_ERROR);
-        $this->assertCount(count($previousErrors), $previous);
-        foreach ($previous as $index => $error) {
-            $this->assertSame([
-                'errorType',
-                'errorMessage',
-                'stack',
-            ], array_keys($error));
-            $this->assertEquals($previousErrors[$index]['errorClass'], $error['errorType']);
-            $this->assertEquals($previousErrors[$index]['errorMessage'], $error['errorMessage']);
-            $this->assertIsArray($error['stack']);
-        }
     }
 }
